requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.4.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.5
blis==0.4.1
Brlapi==0.6.6
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cmake==3.17.2
command-not-found==0.3
contextlib2==0.6.0.post1
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
dlib==19.19.0
docopt==0.6.2
en-core-web-sm==2.2.5
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
google-api-core==1.17.0
google-auth==1.14.2
google-cloud==0.34.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
httplib2==0.9.2
idna==2.9
image-to-scan==0.0.5
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.2
jupyterlab-server==1.1.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==2.1.1
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
nltk==3.5
notebook==6.0.3
numpy==1.18.4
oauth==1.0.1
olefile==0.45.1
opencv-python==4.2.0.34
opencv-python-headless==4.2.0.34
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.6.1
pygobject==3.26.1
pylint==2.5.2
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.16.0
pytesseract==0.3.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.1
regex==2020.5.7
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
rsa==4.0
schema==0.7.2
scikit-learn==0.22.2.post1
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.14.0
sklearn==0.0
soupsieve==2.0
spacy==2.2.4
SpeechRecognition==3.8.1
srsly==1.0.2
system-service==0.3
systemd-python==234
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
toml==0.10.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.3
wadllib==1.3.2
wasabi==0.6.0
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
